requirements.txt
sqlalchemy<2.0
requests
psycopg2-binary
sqlalchemy-redshift
redshift_connector
python-dotenv
schedule
